Remove references to minishell
Every functions/data structures/files begining with the prefix "ms"
(which designated minishell) has a new "sh" prefix (for shell). The name
is generic so that theses files could potentially be used in a 42sh or
else. Now the name of the shell is a macro in the sh_data header.