Tokenize input with new tokens
This includes every new token needed for 21sh (precisely for the
implementation of the grammar described in only_rr), and the "&&"
and "||" tokens, as well as a "&" wich serve as an equivalent of
";" for the moment. A debug file has been added for printing the
resulting of the lexing and of the parsing, when it will be
implemented.