Build parse tree from input
There is also a function to print it, but it should be ehanced.
There is also the fact that the accept_prod and expect_prod are
the exact same function, this will have to be streamlined. The
final has some useless nodes that could be deleted, like TERM,
that could directly be the last one. This not too bad though.