Add the exec_tree structure
This structure is built from the AST (parse_tree), and is directly
used for execution by the execution part of the shell. So the pipes
appear to be working for now. The redirections are not done yet
though, and there is still a lot of fixing and testing to do.