Handle SIGSTOP and SIGSTP
This is very cool. The shell can be interrupted and put back to the
foreground with a sigcont signal. This allows for pausing the program
and doing some shit, but also for lldb (and gdb) to do its job.

lldb: execute ./21sh normally (compiled with -g), then, in an other
terminal, launch lldb. Use the command 'process attach --name 21sh' or
do it via '--pid', then 21sh is stopped while reading and breakpoints
can be set and other funny stuff can be done. Then just type c
(continue), which sends a SIGCONT to 21sh. There you go !